% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_salary_ids.R
\name{add_salary_ids}
\alias{add_salary_ids}
\title{Function to add salary identification}
\usage{
add_salary_ids(projections, salaries)
}
\arguments{
\item{projections}{The unique player projections}

\item{salaries}{player salaries to be appended}
}
\value{
tbl containing players and salaries appended
}
\description{
Function to add salary identification
}
