% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_projections.R
\name{augment_projections}
\alias{augment_projections}
\title{Function to augment projections}
\usage{
augment_projections(projection_data, projection_to_use, value_projection,
  game_style = "classic")
}
\arguments{
\item{projection_data}{Data containing the projections}

\item{projection_to_use}{Name of column containing projection data to use in lineup creation?}

\item{value_projection}{Name of column used to project... value...?}

\item{game_style}{Classic or PickEm (Default "classic")}
}
\value{
list of augmented projections
}
\description{
Function to augment projections
}
