% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outcomes_to_unique_lineups.R,
%   R/augment_unique_lineups.R
\name{augment_unique_lineups}
\alias{augment_unique_lineups}
\alias{augment_unique_lineups}
\title{Function to add outcomes to the unique lineups}
\usage{
augment_unique_lineups(unique_lineup_object, outcome_object, salary_object)

augment_unique_lineups(unique_lineup_object, outcome_object, salary_object)
}
\arguments{
\item{unique_lineup_object}{An object containing unique outcomes}

\item{outcome_object}{the non-unique outcome objects?}

\item{salary_object}{object containing salary information}

\item{unique_lineup_object}{Object containing unique lineups}

\item{outcome_object}{the outcome data}

\item{salary_object}{Object containing salary information}
}
\value{
Data.frame containing augmented lineups

data frame containing augmented outputs
}
\description{
Function to add outcomes to the unique lineups

Adds salary information to lineup objects
}
