% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_lineups.R
\name{export_lineups}
\alias{export_lineups}
\title{Function to save outputs}
\usage{
export_lineups(lineups, entries, site = "draftkings", sport = "nba",
  randomize_entries = TRUE, file_name = paste0(site, "_", sport, "_",
  gsub("[^[:alnum:]]", "", Sys.time()), ".csv"))
}
\arguments{
\item{lineups}{Lineups to save}

\item{entries}{Entry data used to convert lineups to proper format}

\item{site}{The site where lineups will be used. Used to inform formatting}

\item{sport}{The sport of choice}

\item{randomize_entries}{Logical. Indicates whether to randomize order entry.}

\item{file_name}{The name of the file (If not a full path, it is saved to the current directory)}
}
\value{
Logical indicating success
}
\description{
Function to save outputs
}
