% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_nba_names.R
\name{fix_nba_names}
\alias{fix_nba_names}
\title{Function to fix NBA names}
\usage{
fix_nba_names(salary_lookup, projection_list)
}
\arguments{
\item{salary_lookup}{Salary lookup table from make_salary_lu}

\item{projection_list}{list of projections}
}
\description{
Function to fix NBA names
}
