% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_entries.R
\name{import_entries}
\alias{import_entries}
\title{Importing Entry data}
\usage{
import_entries(entries_path, site = "draftkings", csv = TRUE,
  cleansed = TRUE)
}
\arguments{
\item{entries_path}{Path to Entry data}

\item{site}{Betting site. Default is "draftkings"}

\item{csv}{Logical. Not currently used.}

\item{cleansed}{Logical. Also not currently used.}
}
\description{
Importing Entry data
}
