% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_projections.R
\name{import_projections}
\alias{import_projections}
\title{Import projections}
\usage{
import_projections(source = "rotogrinders", site = "draftkings",
  sport = "nba")
}
\arguments{
\item{source}{The site where projections are gathered.  Currently defaults to "rotogrinders" but is not used.}

\item{site}{Betting site. Default is "draftkings"}

\item{sport}{Sport for which lineups are generated. Default = "nba"}
}
\value{
imported projections
}
\description{
Import projections
}
