% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_salaries.R
\name{import_salaries}
\alias{import_salaries}
\title{Importing Salary data}
\usage{
import_salaries(sal_path, site = "draftkings", csv = TRUE,
  from_entry = FALSE)
}
\arguments{
\item{sal_path}{Path to Entry data}

\item{site}{Betting site. Default is "draftkings"}

\item{csv}{Logical. Not currently used.}

\item{cleansed}{Logical. Also not currently used.}
}
\description{
Importing Salary data
}
