% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_possible_lineups.R
\name{make_possible_lineups}
\alias{make_possible_lineups}
\title{Create the possible lineups}
\usage{
make_possible_lineups(player_position_list, salary_cap = 50000,
  salary_min = 30000)
}
\arguments{
\item{player_position_list}{The list of players and their positions}

\item{salary_cap}{Upper limit for spend on a given lineup. Default is 50000}

\item{salary_min}{Lower limit for spend on a given lineup. Default is 30000}
}
\value{
TBL of possible lineups
}
\description{
Create the possible lineups
}
\details{
TODO: remove hard coding of positions
}
