% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_salary_lu.R
\name{make_salary_lu}
\alias{make_salary_lu}
\title{Create the salary lookup table}
\usage{
make_salary_lu(salary_import, site = "draftkings", game_style = "classic",
  from_entry = FALSE)
}
\arguments{
\item{salary_import}{Salary data, generally obtained via \code{import_salaries}}

\item{site}{Betting site. Default is "draftkings"}

\item{game_style}{One of 'classic' or 'pickem'}

\item{from_entry}{Logical. Indicates whether salary_import was generated via Entry file or Salary file from DK.}
}
\description{
Create the salary lookup table
}
