% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_lineups.R
\name{make_unique_lineups}
\alias{make_unique_lineups}
\title{Get the unique lineups from the possible lineup creator}
\usage{
make_unique_lineups(all_possible_lineups)
}
\arguments{
\item{all_possible_lineups}{All the lineups generated by \code{make_possible_lineups}}
}
\description{
Get the unique lineups from the possible lineup creator
}
