% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_lineups.R
\name{optimize_lineups}
\alias{optimize_lineups}
\title{Optimize the lineups}
\usage{
optimize_lineups(unique_lineup_object, position_list = c("pg", "sg", "pf",
  "sf", "c", "g", "f", "util"), n_lineups = 20, max_exposure = 0.65,
  limit_search = 5000, return_freq_table = TRUE, verbose = FALSE)
}
\arguments{
\item{unique_lineup_object}{List of unique objects from \code{make_unique_lineups}}

\item{position_list}{Character vector containing the positions used in the lineups}

\item{n_lineups}{The number of lineups to save (default = 20)}

\item{max_exposure}{The max percentage of lineups in which a single player can be included (default = 65)}

\item{limit_search}{A value used to reduce search space (default 5000)}

\item{return_freq_table}{Logical indicating whether to return the frequency table (default = TRUE). Not currently in use}

\item{verbose}{Logical. Be chatty or not.  Not currently used, so everything is very, very quiet.}
}
\value{
List of optimal lineups
}
\description{
Optimize the lineups
}
