% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_projections.R
\name{scope_projections}
\alias{scope_projections}
\title{Scoping projections}
\usage{
scope_projections(projections)
}
\arguments{
\item{projections}{projection data}
}
\description{
Identifies projections that do not have salary information, and labels them out of scope.
}
