% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ by_position.R
\name{split_by_position}
\alias{split_by_position}
\title{Function to split out be position}
\usage{
split_by_position(projections)
}
\arguments{
\item{projections}{Projection data}
}
\description{
Function to split out be position
}
